test/web-platform-tests/to-upstream/cssom/getComputedStyle-visibility-inheritance-inline.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Computed visibility is inherited for inline styles</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!-- Test for https://github.com/jsdom/jsdom/issues/2865 -->

<body>
  <div id="parent" style="visibility: hidden;">
      <div id="child">Hello, Dave!</div>
  </div>
</body>

<script>
"use strict";

test(() => {
  const element = document.querySelector("#parent");
  assert_equals(getComputedStyle(element).visibility, "hidden");
}, "Sanity check: getComputedStyle returns apparent visibility for inline styles");

test(() => {
  const element = document.querySelector("#child");
  assert_equals(getComputedStyle(element).visibility, "hidden");
}, "Children inherit visibility from their parents inline styles");
</script>
